airflow/requirements.txt
astronomer-cosmos==1.1.3
dbt-bigquery==1.6.5